Add files via upload
We have tested the different pricing engines and compared the time they need to price an european option. We have also changed the time steps. Then our goal was to change the volatility and risk free rate and make them time dependant, but we didn't find how to do it.
In order to cache the common results we thought about adding a vector storage unit, to store them and reuse them without too many computations.